# docs/pro/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Game Panel PRO API",
    "version": "2.0.59",
    "description": "Scoped API. UUIDs remain canonical; server paths also accept global numbers. Current owner privileges are rechecked per request. Native provisioning creates stopped services with central IDs, published templates and per-token budgets. Unknown outcomes are reconciled without automatic re-dispatch. See API-V1-GUIDE.md."
  },
  "security": [
    {
      "apiToken": []
    }
  ],
  "paths": {
    "/api/v1/servers": {
      "get": {
        "operationId": "listServers",
        "summary": "List accessible servers (servers.read)",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "enum": [
                "id",
                "number",
                "name",
                "address"
              ],
              "default": "id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          },
          {
            "name": "game",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          },
          {
            "name": "node",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Page ordered by global server ID",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "nextCursor",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Server"
                      }
                    },
                    "nextCursor": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "description": "Cached fleet metadata, no per-row node requests. Default UUID ordering retained; address uses numeric IP and port ordering. Reuse filters with after; restart on invalid_cursor. stale=true forces status=unknown."
      },
      "post": {
        "operationId": "installServer",
        "summary": "Install a stopped Native server",
        "description": "Administrator plus provisioning policy. Durable admission. Same request/key returns same operation; changed payload gives 409. Budget includes uncertain/failed admissions. Token gains access to its created server. Location points to progress. 409 data.status=uncertain means poll, never submit a new key. Installation keys are separate from backup/power.",
        "x-scopes": [
          "servers.create"
        ],
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/ProvisionAdmission"
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "description": "Conflict error, or unconfirmed dispatch with data: ProvisionAdmission (status=uncertain) and Location.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProvisionRequest"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]{16,128}$"
            }
          }
        ]
      }
    },
    "/api/v1/servers/{id}": {
      "get": {
        "operationId": "getServer",
        "summary": "Read a scoped server (servers.read)",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Fleet UUID or global server number, e.g. 101"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Server inventory record",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Server"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "description": "Live runtime detail when reachable, cached inventory with available=false otherwise. No environment, commands, host paths or credentials. Resource sample requires resources.read. Unsupported or missing values are null."
      }
    },
...
    "/api/v1/operations/{id}": {
      "get": {
        "operationId": "getOperation",
        "summary": "Read a backup operation (operations.read)",
        "description": "Requires the same owner, current backups.create permission and a token scoped to this server. Runtime replacement returns 409; offline runtime returns 503 instead of an invented result. Internal output, paths and runtime IDs are omitted.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          },
          "200": {
            "description": "Current runtime operation status or explicit uncertain dispatch",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data",
                    "requestId"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Operation"
                    },
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v1/templates": {
      "get": {
        "operationId": "listPublishedTemplates",
        "summary": "Published Native templates",
        "description": "Administrator only. Filtered by token policy. Version, name, game, architectures, public port definitions and variables. Scripts, images, mounts and secret defaults are omitted.",
        "x-scopes": [
          "templates.read"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/nodes": {
      "get": {
        "operationId": "listProvisionNodes",
        "summary": "Enabled node choices",
        "description": "Administrator only. Filtered by token policy. No host origins or credentials.",
        "x-scopes": [
          "nodes.read"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "location": {
                            "type": [
                              "string",
                              "null"
                            ]
                          }
                        }
                      }
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/nodes/{id}/allocations": {
      "get": {
        "operationId": "listNodeAllocations",
        "summary": "Configured game IP and port pools",
        "description": "Configured pools, not reservations. Administrator only. Plan and install check available ports.",
        "x-scopes": [
          "nodes.read"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "properties": {
                          "ip": {
                            "type": "string"
                          },
                          "tcp": {
                            "type": "string"
                          },
                          "udp": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "requestId": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          },
          "429": {
            "$ref": "#/components/responses/Error"
          },
          "503": {
            "$ref": "#/components/responses/Error"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "gpp_ token"
      }
    },
    "schemas": {
...
      "ProvisionAdmission": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "const": "server.install"
          },
          "status": {
            "enum": [
              "dispatching",
              "accepted",
              "uncertain"
            ]
          },
          "serverId": {
            "type": [
              "string",
              "null"
            ]
          },
          "startAfterInstall": {
            "const": false
          }
        }
      },
...
    },
    "responses": {
      "Error": {
        "description": "Structured error; requestId also appears in X-Request-ID",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "error",
                "requestId"
              ],
              "properties": {
                "error": {
                  "type": "object",
                  "required": [
                    "code",
                    "message"
                  ],
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "requestId": {
                  "type": "string",
                  "format": "uuid"
                }
              }
            }
          }
        }
      }
    }
  }
}